requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-groq
langchain-text-splitters
langchain-chroma
streamlit
duckduckgo-search
arxiv
wikipedia
bs4
validators
pymupdf
pymupdf
unstructured
youtube_transcript_api
pytube
chromadb
sentence_transformers
huggingface_hub
pandas
numpy
SQLAlchemy
mysql-connector-python
numexpr
python-dotenv
ipykernel
pydantic
